Update README.md
Welcome to the Stopwatch repository built using Java Swing This project offers an exciting implementation of the stopwatch.The Stopwatch Application is a software-based timekeeping tool designed to accurately measure and display elapsed time. It provides users with the ability to start, stop, and reset the timer, as well as record split times for tracking specific intervals.